v.0.5

+++ rush00/inc/Vec2i.hpp
@@ -13,21 +13,21 @@ class Vec2i
 public:
 	/* Canonical Form */
 	Vec2i();
-	Vec2i(int_fast8_t x, int_fast8_t y);
+	Vec2i(int_fast16_t x, int_fast16_t y);
 	Vec2i(Vec2i const &src);
 	~Vec2i();
 	Vec2i &operator=(Vec2i const &rhs);
 
-	int_fast8_t getX() const;
-	int_fast8_t getY() const;
+	int_fast16_t getX() const;
+	int_fast16_t getY() const;
 
-	void setX(int_fast8_t x);
-	void setY(int_fast8_t y);
-	void setXY(int_fast8_t x, int_fast8_t y);
+	void setX(int_fast16_t x);
+	void setY(int_fast16_t y);
+	void setXY(int_fast16_t x, int_fast16_t y);
 
 private:
-	int_fast8_t _x;
-	int_fast8_t _y;
+	int_fast16_t _x;
+	int_fast16_t _y;
 
 };
