+++ rush00/inc/Vec2ui.hpp
@@ -12,21 +12,21 @@ class Vec2ui
 public:
 	/* Canonical Form */
 	Vec2ui();
-	Vec2ui(uint_fast8_t x, uint_fast8_t y);
+	Vec2ui(uint_fast16_t x, uint_fast16_t y);
 	Vec2ui(Vec2ui const &src);
 	~Vec2ui();
 	Vec2ui &operator=(Vec2ui const &rhs);
 
-	uint_fast8_t getX() const;
-	uint_fast8_t getY() const;
+	uint_fast16_t getX() const;
+	uint_fast16_t getY() const;
 
-	void setX(uint_fast8_t x);
-	void setY(uint_fast8_t y);
-	void setXY(uint_fast8_t x, uint_fast8_t y);
+	void setX(uint_fast16_t x);
+	void setY(uint_fast16_t y);
+	void setXY(uint_fast16_t x, uint_fast16_t y);
 
 private:
-	uint_fast8_t _x;
-	uint_fast8_t _y;
+	uint_fast16_t _x;
+	uint_fast16_t _y;
 
 };
